requirements.txt
openiap-edge==0.0.36
beautifulsoup4==4.12.3
requests==2.32.3